refactor(ai): split Data Machine tool registry source
